requirements.txt
flask
requests
pandas
matplotlib
python-pptx